PLAT-42: Added createTimeStamp field, sqs service impl usage
