+++ ascent-document-service/src/main/java/gov/va/ascent/document/service/rest/provider/DocumentServiceEndPoint.java
@@ -23,9 +23,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import gov.va.ascent.document.service.api.DocumentService;
 import gov.va.ascent.document.service.api.transfer.GetDocumentTypesResponse;
-import gov.va.ascent.document.sqs.service.QueueService;
 import gov.va.ascent.framework.swagger.SwaggerResponseMessages;
 import gov.va.ascent.starter.aws.autoconfigure.s3.services.S3Services;
+import gov.va.ascent.starter.aws.autoconfigure.sqs.services.SQSServices;
 import io.swagger.annotations.ApiOperation;
 
 @RestController
@@ -41,7 +41,7 @@ public class DocumentServiceEndPoint implements HealthIndicator, SwaggerResponse
 	S3Services s3Services;
 	    
 	@Autowired
-	QueueService queueService;
+	SQSServices sqsServices;
 	
 	@Value("${ascent.s3.uploadfile}")
 	private String uploadFilePath;
@@ -79,7 +79,7 @@ public class DocumentServiceEndPoint implements HealthIndicator, SwaggerResponse
 		LOGGER.info("Sending message {}.", "Sample Test Message");
 		
 		String jsonMessage = documentService.getMessageAttributes(documentOne.getOriginalFilename());
-		queueService.sendMessage(jsonMessage);
+		sqsServices.sendMessage(jsonMessage);
 		return ResponseEntity.ok().build();
     }	
 	
@@ -87,7 +87,7 @@ public class DocumentServiceEndPoint implements HealthIndicator, SwaggerResponse
 	public ResponseEntity<?> sendMessage(@RequestBody String message) {
 	   LOGGER.info("Sending message {}.", message);
 	   String jsonMessage = documentService.getMessageAttributes(message);
-	   queueService.sendMessage(jsonMessage);
+	   sqsServices.sendMessage(jsonMessage);
 	   return ResponseEntity.ok().build();
 	}
