--- ascent-document-service/src/main/java/gov/va/ascent/document/sqs/service/QueueService.java
@@ -1,12 +1,0 @@
-package gov.va.ascent.document.sqs.service;
-
-import org.springframework.http.ResponseEntity;
-
-public interface QueueService {
-	/**
-     * Send a message to SQS
-     * @param request String message
-     * @return ResponseEntity<String> JMS Message ID
-     */
-    public ResponseEntity<String> sendMessage(final String request);
-}
